DOES NOT BUILD
Removed Jade because it's overkill for the simple tasks we want to do.
Rewrote it in html and switched up the project structure. Also quickly created UpdateRequest & Response
classes. I probably did those wrong & that's why its not building "some tests are failing.".

Currently the idea is that there'll be a separate HTML page for each book? Perhaps that's
not what you were thinking, in that case we can change it.

For Scrolling on the Site i'll look into something called Jscroll.

For our search function, if we want to implement it. Postgres has
builtin functions for fuzzy search. We just need to tell Postgres
to index the DB, i'll read up more on that later.

Perhaps you could take a look in the BookResource & my Update/Response classes
and fix what i did wrong if you have time! :)